requirements.txt
flask
flask-socketio
eventlet
Pillow